requirements.txt
streamlit==0.80.0
Pillow==8.2.0
requests-toolbelt==0.9.1